Hotfix

+++ src/main/java/com/EIPplatform/model/dto/report/report05/airemmissionmanagement/airautomonitoringstat/AirAutoMonitoringStatDTO.java
@@ -22,8 +22,8 @@ public class AirAutoMonitoringStatDTO {
 
     Integer valError;
 
-    BigDecimal ratioReceivedDesign;
+    Double ratioReceivedDesign;
 
-    BigDecimal ratioErrorReceived;
+    Double ratioErrorReceived;
 
 }
