+++ src/main/java/com/EIPplatform/model/dto/report/report05/airemmissionmanagement/airautoqcvnexceedance/AirAutoQcvnExceedanceDTO.java
@@ -20,5 +20,5 @@ public class AirAutoQcvnExceedanceDTO {
 
     Integer qcvnLimitValue;
 
-    BigDecimal exceedRatioPercent;
+    Double exceedRatioPercent;
 }
