+++ src/main/java/com/EIPplatform/model/dto/report/report05/airemmissionmanagement/airmonitoringexceedance/AirMonitoringExceedanceDTO.java
@@ -27,8 +27,8 @@ public class AirMonitoringExceedanceDTO {
 
     String exceededParam;
 
-    BigDecimal resultValue;
+    Double resultValue;
 
-    BigDecimal qcvnLimit;
+    Double qcvnLimit;
 
 }
