+++ src/main/java/com/EIPplatform/model/dto/report/report05/wastemanagement/popinventorystat/PopInventoryStatDTO.java
@@ -16,9 +16,9 @@ public class PopInventoryStatDTO {
     String popName;
     String casCode;
     LocalDate importDate;
-    BigDecimal importVolume;
+    Double importVolume;
     String concentration;
-    BigDecimal volumeUsed;
-    BigDecimal volumeStocked;
+    Double volumeUsed;
+    Double volumeStocked;
     String complianceResult;
 }
