+++ src/main/java/com/EIPplatform/model/dto/report/report05/wastemanagement/selftreatedhwstat/SelfTreatedHwStatDTO.java
@@ -14,6 +14,6 @@ public class SelfTreatedHwStatDTO {
     Long selfTreatedId;
     String wasteName;
     String hwCode;
-    BigDecimal volumeKg;
+    Double volumeKg;
     String selfTreatmentMethod;
 }
